+++ NEWS.md
@@ -1,5 +1,7 @@
 # CHANGES IN markdown VERSION 1.12
 
+- Provided three internal functions `html_document`, `html_vignette`, and `pdf_document` as compatibility layers to functions of the same names in the **rmarkdown** package (thanks, @jangorecki, #108). See [the appendix](https://cran.r-project.org/package=markdown/vignettes/intro.html#for-rmarkdown-users) in `vignette('intro', package = 'markdown')` for details.
+
 - The default HTML template no longer wraps meta variables `include-before` and `include-after` inside `<div></div>`, because their values may contain incomplete HTML tags, e.g., `include-before = '<div>'` and `include-after = '</div>'`.
 
 # CHANGES IN markdown VERSION 1.11